requirements.txt
pyyaml
tqdm
colorama
watchdog
customtkinter
pillow
pyinstaller
